Move password-related functions into a dedicated file
